Created class DropFlow that keeps track of a Drops current state. It's provied to the Stream Stones serveFailure, serveErr and finish
The tests of module river werent reliable as a test Suite was
signaled that the topicCb is done and the vows can commence
before River was actually completely done with the Drop. This
has been resolved by providing River with a cb that is called
once River is done handling a Drop. The cb is only to be used
for testing purposes. Now the test Suite proceeds with the vows
only when River signals it is done with the Drop and an http
response is received by the test Suite if such response is
expected

drop.js: Now creates a DropFlow that is used to keep track of the
Drop's progress

main.js: Exposed DropFlow

river.js: River() now accepts a test cb to call once handling of
a Drop is done

river.js: Various args dont need to be passed down to RiverDrop
anymore (thanks to DropFlow)

riverdrop.js: Various args dont need to be passed down to Drop
anymore (thanks to DropFlow)

stone.js: Changed a FuncVer

stoneservefailure.js: Now gives the DropFlow to the Stream Stone

stoneserveerr.js: Now gives the DropFlow to the Stream Stone

stonefinish.js: Now gives the DropFlow to the Stream Stone

stonevalidate.js: Minor modification (added spacebar)

stream.js: Started using DropFlow

river-test.js: Made tests more reliable as Suite Vows no longer
commence until the River is completely done with the Drop and any
expected http response has been received

river-test.js: Incorporated DropFlow as it had to be added to
many func headers. Also added tests that test the functionality
of DropFlow in the Stream Stones that use it